+++ lib/ShapedJson/shaped-json.c
@@ -1587,7 +1587,7 @@ static bool StringifyJsonShapeDataArray (TRI_shaper_t* shaper,
                                          char const* data,
                                          uint64_t size,
                                          bool braces,
-                                         size_t* num) {
+                                         uint64_t* num) {
   TRI_array_shape_t const* s;
   TRI_shape_aid_t const* aids;
   TRI_shape_sid_t const* sids;
@@ -2229,7 +2229,7 @@ bool TRI_StringifyAugmentedShapedJson (TRI_shaper_t* shaper,
                                        TRI_json_t const* augment) {
   TRI_shape_t const* shape;
   bool ok;
-  size_t num;
+  uint64_t num;
   int res;
 
   shape = shaper->lookupShapeId(shaper, shaped->_sid);
